cleanup: optimize server logic
