+++ fss/.env.example
@@ -3,9 +3,11 @@ APP_NAME=Fss
 APP_DESC="Fss aims to one of the best scaffolding in the PyWeb field."
 VERSION=1.0.0-beta.1
 MODE=dev
+HOST=0.0.0.0
 PORT=9010
 WORKERS=1
 API_VERSION=/api/v1
+ECHO_SQL=True
 
 # Time zone
 WIN_TZ="China Standard Time"
